Properly document all functions, including a type for the `systemData` return value
