+++ modules/exoplanets/CMakeLists.txt
@@ -25,6 +25,7 @@
 include(${PROJECT_SOURCE_DIR}/support/cmake/module_definition.cmake)
 
 set(HEADER_FILES
+    datastructure.h    
     exoplanetshelper.h
     exoplanetsmodule.h
     rendering/renderableorbitdisc.h
@@ -33,6 +34,7 @@ set(HEADER_FILES
 source_group("Header Files" FILES ${HEADER_FILES})
 
 set(SOURCE_FILES
+    datastructure.cpp    
     exoplanetshelper.cpp
     exoplanetsmodule.cpp
     exoplanetsmodule_lua.inl
