+++ modules/exoplanets/exoplanetsmodule.cpp
@@ -24,6 +24,7 @@
 
 #include <modules/exoplanets/exoplanetsmodule.h>
 
+#include <modules/exoplanets/datastructure.h>
 #include <modules/exoplanets/exoplanetshelper.h>
 #include <modules/exoplanets/rendering/renderableorbitdisc.h>
 #include <modules/exoplanets/tasks/exoplanetsdatapreparationtask.h>
@@ -407,7 +408,8 @@ void ExoplanetsModule::internalInitialize(const ghoul::Dictionary& dict) {
 std::vector<documentation::Documentation> ExoplanetsModule::documentations() const {
     return {
         ExoplanetsDataPreparationTask::documentation(),
-        RenderableOrbitDisc::Documentation()
+        RenderableOrbitDisc::Documentation(),
+        ExoplanetSystem::Documentation()
     };
 }
