+++ modules/exoplanets/tasks/exoplanetsdatapreparationtask.h
@@ -27,7 +27,7 @@
 
 #include <openspace/util/task.h>
 
-#include <modules/exoplanets/exoplanetshelper.h>
+#include <modules/exoplanets/datastructure.h>
 
 #include <openspace/properties/vector/vec3property.h>
 #include <filesystem>
